Ring join perfected. Wait for token still not working

+++ client/communication/TokenMessage.java
@@ -9,7 +9,7 @@ import distributed.PeerManager;
 
 @XmlRootElement
 public class TokenMessage extends Message {
-	public int counter;
+	public long counter;
 	public int iddo;
 	
 	public TokenMessage(){
